fix: register native routines
We used to get a NOTE during `devtools::check`. This commit fixes it.

As a reference, the NOTE was:

```
File ‘mgarchBEKK/libs/mgarchBEKK.so’:
  Found no calls to: ‘R_registerRoutines’, ‘R_useDynamicSymbols’

It is good practice to register native routines and to disable symbol
search.

See ‘Writing portable packages’ in the ‘Writing R Extensions’ manual.
```

Following resources helped:

- https://ironholds.org/registering-routines/
- https://stackoverflow.com/questions/42313373/r-cmd-check-note-found-no-calls-to-r-registerroutines-r-usedynamicsymbols
- https://stackoverflow.com/questions/8407615/adding-usedynlib-through-roxygen